Fix bags with length of description in yalObjectType and fix nulPointerException in ageRange.dat.

+++ src/main/java/ru/yalrb/entity/models/AgeRange.java
@@ -16,7 +16,7 @@ import lombok.RequiredArgsConstructor;
 public class AgeRange {
     @Id
     @GeneratedValue
-    private Byte id;
+    private Long id;
 
     private Integer fromAge;
